Move App Doctor evidence types out of the rules layer
Scanners own discovery types so rules depend on evidence, not the reverse.
